Base extended DB schema
