+++ accounting/endpoints.py
@@ -7,10 +7,7 @@ from .schemas import (
     UserCreate,
     UserPasswordChange
 )
-from fastapi.encoders import jsonable_encoder
-from fastapi import Depends, Request, Response
-from fastapi.security import OAuth2PasswordRequestForm
-from .auth import create_access_token, get_user_by_token
+from fastapi import Request, Response
 
 async def get_all_users(request: Request, offset: int = 0, limit: int = 100):
     """
@@ -95,12 +92,3 @@ async def delete_user(id: str):
     await User.delete_by_id(id)
     return Response(status_code=204)
 
-
-async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
-    user = await User.authenticate_user(form_data.username,form_data.password)
-    access_token = create_access_token(
-        data={"sub": user.username})
-    return {"access_token": access_token, "token_type": "bearer"}
-
-async def get_current_user(current_user: User = Depends(get_user_by_token)):
-    return current_user
