--- accounting/piccolo_migrations/2022-07-03T16-25-35-600584.py
@@ -1,397 +1,0 @@
-from piccolo.apps.migrations.auto.migration_manager import MigrationManager
-from piccolo.columns.column_types import Boolean
-from piccolo.columns.column_types import Serial
-from piccolo.columns.column_types import Text
-from piccolo.columns.column_types import Timestamp
-from piccolo.columns.column_types import Timestamptz
-from piccolo.columns.defaults.timestamp import TimestampOffset
-from piccolo.columns.defaults.timestamptz import TimestamptzCustom
-from piccolo.columns.defaults.timestamptz import TimestamptzNow
-from piccolo.columns.indexes import IndexMethod
-
-
-ID = "2022-07-03T16:25:35:600584"
-VERSION = "0.80.0"
-DESCRIPTION = ""
-
-
-async def forwards():
-    manager = MigrationManager(
-        migration_id=ID, app_name="user", description=DESCRIPTION
-    )
-
-    manager.add_table("User", tablename="users")
-
-    manager.add_table("Sessions", tablename="sessions")
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="id",
-        db_column_name="id",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": True,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="username",
-        db_column_name="username",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": True,
-            "index": True,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="email",
-        db_column_name="email",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "nullable": True,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="password",
-        db_column_name="password",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="first_name",
-        db_column_name="first_name",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": True,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="last_name",
-        db_column_name="last_name",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": True,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="active",
-        db_column_name="active",
-        column_class_name="Boolean",
-        column_class=Boolean,
-        params={
-            "nullable": False,
-            "default": True,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="admin",
-        db_column_name="admin",
-        column_class_name="Boolean",
-        column_class=Boolean,
-        params={
-            "default": False,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="superuser",
-        db_column_name="superuser",
-        column_class_name="Boolean",
-        column_class=Boolean,
-        params={
-            "default": False,
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="created_at",
-        db_column_name="created_at",
-        column_class_name="Timestamptz",
-        column_class=Timestamptz,
-        params={
-            "nullable": False,
-            "default": TimestamptzCustom(
-                year=2022, month=7, day=7, hour=16, second=35, microsecond=588373
-            ),
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="updated_at",
-        db_column_name="updated_at",
-        column_class_name="Timestamptz",
-        column_class=Timestamptz,
-        params={
-            "nullable": False,
-            "default": TimestamptzCustom(
-                year=2022, month=7, day=7, hour=16, second=35, microsecond=588544
-            ),
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="last_login",
-        db_column_name="last_login",
-        column_class_name="Timestamptz",
-        column_class=Timestamptz,
-        params={
-            "nullable": True,
-            "default": TimestamptzNow(),
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="User",
-        tablename="users",
-        column_name="birthdate",
-        db_column_name="birthdate",
-        column_class_name="Timestamptz",
-        column_class=Timestamptz,
-        params={
-            "nullable": True,
-            "default": TimestamptzNow(),
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Sessions",
-        tablename="sessions",
-        column_name="token",
-        db_column_name="token",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "length": 100,
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Sessions",
-        tablename="sessions",
-        column_name="user_id",
-        db_column_name="user_id",
-        column_class_name="Text",
-        column_class=Text,
-        params={
-            "default": "",
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Sessions",
-        tablename="sessions",
-        column_name="expiry_date",
-        db_column_name="expiry_date",
-        column_class_name="Timestamp",
-        column_class=Timestamp,
-        params={
-            "default": TimestampOffset(days=0, hours=1, minutes=0, seconds=0),
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Sessions",
-        tablename="sessions",
-        column_name="max_expiry_date",
-        db_column_name="max_expiry_date",
-        column_class_name="Timestamp",
-        column_class=Timestamp,
-        params={
-            "default": TimestampOffset(days=7, hours=0, minutes=0, seconds=0),
-            "null": False,
-            "primary_key": False,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": None,
-            "secret": False,
-        },
-    )
-
-    manager.add_column(
-        table_class_name="Sessions",
-        tablename="sessions",
-        column_name="id",
-        db_column_name="id",
-        column_class_name="Serial",
-        column_class=Serial,
-        params={
-            "null": False,
-            "primary_key": True,
-            "unique": False,
-            "index": False,
-            "index_method": IndexMethod.btree,
-            "choices": None,
-            "db_column_name": "id",
-            "secret": False,
-        },
-    )
-
-    return manager
