+++ accounting/piccolo_migrations/2022-07-15T14-58-35-521868.py
@@ -1,0 +1,828 @@
+from piccolo.apps.migrations.auto.migration_manager import MigrationManager
+from piccolo.columns.base import OnDelete
+from piccolo.columns.base import OnUpdate
+from piccolo.columns.column_types import Boolean
+from piccolo.columns.column_types import ForeignKey
+from piccolo.columns.column_types import Serial
+from piccolo.columns.column_types import Text
+from piccolo.columns.column_types import Timestamp
+from piccolo.columns.defaults.timestamp import TimestampCustom
+from piccolo.columns.defaults.timestamp import TimestampNow
+from piccolo.columns.defaults.timestamp import TimestampOffset
+from piccolo.columns.indexes import IndexMethod
+from piccolo.table import Table
+
+
+class Group(Table, tablename="groups"):
+    id = Text(
+        default="",
+        null=False,
+        primary_key=True,
+        unique=False,
+        index=True,
+        index_method=IndexMethod.btree,
+        choices=None,
+        db_column_name=None,
+        secret=False,
+    )
+
+
+class Permission(Table, tablename="permissions"):
+    id = Text(
+        default="",
+        null=False,
+        primary_key=True,
+        unique=False,
+        index=True,
+        index_method=IndexMethod.btree,
+        choices=None,
+        db_column_name=None,
+        secret=False,
+    )
+
+
+class Role(Table, tablename="roles"):
+    id = Text(
+        default="",
+        null=False,
+        primary_key=True,
+        unique=False,
+        index=True,
+        index_method=IndexMethod.btree,
+        choices=None,
+        db_column_name=None,
+        secret=False,
+    )
+
+
+class User(Table, tablename="users"):
+    id = Text(
+        default="",
+        null=False,
+        primary_key=True,
+        unique=False,
+        index=True,
+        index_method=IndexMethod.btree,
+        choices=None,
+        db_column_name=None,
+        secret=False,
+    )
+
+
+ID = "2022-07-15T14:58:35:521868"
+VERSION = "0.80.0"
+DESCRIPTION = ""
+
+
+async def forwards():
+    manager = MigrationManager(
+        migration_id=ID, app_name="accounting", description=DESCRIPTION
+    )
+
+    manager.add_table("Group", tablename="groups")
+
+    manager.add_table("User", tablename="users")
+
+    manager.add_table("Sessions", tablename="sessions")
+
+    manager.add_table("Permission", tablename="permissions")
+
+    manager.add_table("M2MUserGroup", tablename="m2_m_user_group")
+
+    manager.add_table("M2MUserRole", tablename="m2_m_user_role")
+
+    manager.add_table("Role", tablename="roles")
+
+    manager.add_table("Policy", tablename="policies")
+
+    manager.add_column(
+        table_class_name="Group",
+        tablename="groups",
+        column_name="id",
+        db_column_name="id",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Group",
+        tablename="groups",
+        column_name="name",
+        db_column_name="name",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": True,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Group",
+        tablename="groups",
+        column_name="active",
+        db_column_name="active",
+        column_class_name="Boolean",
+        column_class=Boolean,
+        params={
+            "nullable": False,
+            "default": True,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="id",
+        db_column_name="id",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="username",
+        db_column_name="username",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": True,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="email",
+        db_column_name="email",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "nullable": True,
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="password",
+        db_column_name="password",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="first_name",
+        db_column_name="first_name",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="last_name",
+        db_column_name="last_name",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="active",
+        db_column_name="active",
+        column_class_name="Boolean",
+        column_class=Boolean,
+        params={
+            "nullable": False,
+            "default": True,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="admin",
+        db_column_name="admin",
+        column_class_name="Boolean",
+        column_class=Boolean,
+        params={
+            "default": False,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="superuser",
+        db_column_name="superuser",
+        column_class_name="Boolean",
+        column_class=Boolean,
+        params={
+            "default": False,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="created_at",
+        db_column_name="created_at",
+        column_class_name="Timestamp",
+        column_class=Timestamp,
+        params={
+            "nullable": False,
+            "default": TimestampCustom(
+                year=2022, month=7, day=7, hour=14, second=35, microsecond=509098
+            ),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="updated_at",
+        db_column_name="updated_at",
+        column_class_name="Timestamp",
+        column_class=Timestamp,
+        params={
+            "nullable": False,
+            "default": TimestampCustom(
+                year=2022, month=7, day=7, hour=14, second=35, microsecond=509164
+            ),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="last_login",
+        db_column_name="last_login",
+        column_class_name="Timestamp",
+        column_class=Timestamp,
+        params={
+            "nullable": True,
+            "default": TimestampNow(),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="User",
+        tablename="users",
+        column_name="birthdate",
+        db_column_name="birthdate",
+        column_class_name="Timestamp",
+        column_class=Timestamp,
+        params={
+            "nullable": True,
+            "default": TimestampNow(),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Sessions",
+        tablename="sessions",
+        column_name="token",
+        db_column_name="token",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "length": 100,
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Sessions",
+        tablename="sessions",
+        column_name="user_id",
+        db_column_name="user_id",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Sessions",
+        tablename="sessions",
+        column_name="expiry_date",
+        db_column_name="expiry_date",
+        column_class_name="Timestamp",
+        column_class=Timestamp,
+        params={
+            "default": TimestampOffset(days=0, hours=1, minutes=0, seconds=0),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Sessions",
+        tablename="sessions",
+        column_name="max_expiry_date",
+        db_column_name="max_expiry_date",
+        column_class_name="Timestamp",
+        column_class=Timestamp,
+        params={
+            "default": TimestampOffset(days=7, hours=0, minutes=0, seconds=0),
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Sessions",
+        tablename="sessions",
+        column_name="id",
+        db_column_name="id",
+        column_class_name="Serial",
+        column_class=Serial,
+        params={
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": "id",
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Permission",
+        tablename="permissions",
+        column_name="id",
+        db_column_name="id",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Permission",
+        tablename="permissions",
+        column_name="name",
+        db_column_name="name",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": True,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Permission",
+        tablename="permissions",
+        column_name="object",
+        db_column_name="object",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": True,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Permission",
+        tablename="permissions",
+        column_name="description",
+        db_column_name="description",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="M2MUserGroup",
+        tablename="m2_m_user_group",
+        column_name="user",
+        db_column_name="user",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": User,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "target_column": None,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="M2MUserGroup",
+        tablename="m2_m_user_group",
+        column_name="group",
+        db_column_name="group",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Group,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "target_column": None,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="M2MUserRole",
+        tablename="m2_m_user_role",
+        column_name="user",
+        db_column_name="user",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": User,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "target_column": None,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="M2MUserRole",
+        tablename="m2_m_user_role",
+        column_name="role",
+        db_column_name="role",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Role,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "target_column": None,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Role",
+        tablename="roles",
+        column_name="id",
+        db_column_name="id",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Role",
+        tablename="roles",
+        column_name="name",
+        db_column_name="name",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": False,
+            "unique": True,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Role",
+        tablename="roles",
+        column_name="active",
+        db_column_name="active",
+        column_class_name="Boolean",
+        column_class=Boolean,
+        params={
+            "nullable": False,
+            "default": True,
+            "null": False,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Policy",
+        tablename="policies",
+        column_name="id",
+        db_column_name="id",
+        column_class_name="Text",
+        column_class=Text,
+        params={
+            "default": "",
+            "null": False,
+            "primary_key": True,
+            "unique": False,
+            "index": True,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Policy",
+        tablename="policies",
+        column_name="permission",
+        db_column_name="permission",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Permission,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "target_column": None,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    manager.add_column(
+        table_class_name="Policy",
+        tablename="policies",
+        column_name="role",
+        db_column_name="role",
+        column_class_name="ForeignKey",
+        column_class=ForeignKey,
+        params={
+            "references": Role,
+            "on_delete": OnDelete.cascade,
+            "on_update": OnUpdate.cascade,
+            "target_column": None,
+            "null": True,
+            "primary_key": False,
+            "unique": False,
+            "index": False,
+            "index_method": IndexMethod.btree,
+            "choices": None,
+            "db_column_name": None,
+            "secret": False,
+        },
+    )
+
+    return manager
