+++ accounting/schemas.py
@@ -44,10 +44,3 @@ class UserPasswordChange(BaseModel):
     class Config:
         orm_mode = True
 
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    username: str | None = None
