+++ config.ini
@@ -47,7 +47,7 @@ db_name = fastapi-boilerplate
 db_username = fastapi-boilerplate
 db_password = fastapi-boilerplate
 
-db_vault_enable = 1
+db_vault_enable = 0
 db_vault_role = testrole02
 db_vault_static = 1
 db_vault_storage = database
